made all dtypes the same

+++ tests/input/inference_dataset_tests.py
@@ -185,8 +185,8 @@ class TestInferenceDataSet(unittest.TestCase):
         # input stack should be normalized, not target
         self.assertEqual(input_stack.max(), 0.0)
         self.assertEqual(target_stack.max(), 1)
-        self.assertEqual(input_stack.dtype, np.float64)
-        self.assertEqual(target_stack.dtype, np.float64)
+        self.assertEqual(input_stack.dtype, np.float32)
+        self.assertEqual(target_stack.dtype, np.float32)
 
     def test__getitem__regression(self):
         dataset_config = {
@@ -209,6 +209,6 @@ class TestInferenceDataSet(unittest.TestCase):
         # input stack should be normalized, not target
         self.assertEqual(input_stack.max(), 0.0)
         self.assertEqual(target_stack.max(), 0.0)
-        self.assertEqual(input_stack.dtype, np.float64)
-        self.assertEqual(target_stack.dtype, np.float64)
+        self.assertEqual(input_stack.dtype, np.float32)
+        self.assertEqual(target_stack.dtype, np.float32)
